requirements.txt
# requirements.txt
streamlit>=1.30
openai==1.54.3
azure-ai-ml==1.21.1
azure-identity==1.19.0
pypdf==5.1.0
requests==2.32.3
python-dotenv==1.0.1
numpy==1.26.4
PyMuPDF>=1.24.0